switch to react-chartjs-2 & made Linechart component
