+++ package.json
@@ -15,6 +15,7 @@
     "mongoose": "^6.6.7",
     "next": "12.3.1",
     "react": "18.2.0",
+    "react-chartjs-2": "^4.3.1",
     "react-dom": "18.2.0",
     "styled-components": "^5.3.6"
   },
